project restructuring related to targets' source files + nested CMakeLists files inside targets
