+++ Engine/Source/Core/CommandQueue.cpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "stdafx.h"
-#include "Common/ScaldCoreDefines.h"
+#include "ScaldCoreDefines.h"
 #include "CommandQueue.h"
 
 CommandQueue::CommandQueue(const ComPtr<ID3D12Device2>& device, D3D12_COMMAND_LIST_TYPE type)
