+++ Engine/Source/Core/ScaldCoreDefines.h
