+++ Engine/Source/Core/ScaldCoreTypes.h
