+++ Engine/Source/Core/ScaldUtil.cpp
